🚧 WIP Scan Mojo
