+++ src/main/java/com/contrastsecurity/maven/plugin/ContrastInstallAgentMojo.java
@@ -1,21 +1,61 @@
 package com.contrastsecurity.maven.plugin;
 
+import com.contrastsecurity.exceptions.UnauthorizedException;
+import com.contrastsecurity.models.AgentType;
+import com.contrastsecurity.models.Applications;
 import com.contrastsecurity.sdk.ContrastSDK;
 import java.io.File;
+import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.maven.model.Plugin;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.plugins.annotations.Parameter;
+import org.apache.maven.project.MavenProject;
 
 @Mojo(name = "install", defaultPhase = LifecyclePhase.VALIDATE, requiresOnline = true)
-public class InstallAgentContrastMavenMojo extends AbstractContrastMavenPluginMojo {
+public class ContrastInstallAgentMojo extends AbstractAssessMojo {
+
+  @Parameter(defaultValue = "${project}", readonly = true)
+  private MavenProject project;
+
+  @Parameter(property = "skipArgLine")
+  boolean skipArgLine;
+
+  @Parameter(property = "standalone")
+  boolean standalone;
+
+  @Parameter(property = "profile")
+  private String profile;
+
+  @Parameter(property = "environment")
+  private String environment;
+
+  @Parameter(property = "serverPath")
+  String serverPath;
+
+  @Parameter(property = "jarPath")
+  private String jarPath;
+
+  String contrastAgentLocation;
+
+  @Parameter(property = "applicationSessionMetadata")
+  private String applicationSessionMetadata;
+
+  @Parameter(property = "applicationTags")
+  private String applicationTags;
+
+  @Parameter(property = "appVersion")
+  String appVersion;
+
   String applicationName;
 
   static Map<String, String> environmentToSessionMetadata = new TreeMap<String, String>();
@@ -81,7 +121,36 @@ public class InstallAgentContrastMavenMojo extends AbstractContrastMavenPluginMo
     }
   }
 
-  public String computeAppVersion(Date currentDate) {
+  private String getAppName(ContrastSDK contrastSDK, String applicationId)
+      throws MojoExecutionException {
+    Applications applications;
+    try {
+      applications = contrastSDK.getApplication(orgUuid, applicationId);
+    } catch (Exception e) {
+      String logMessage;
+      if (e.getMessage().contains("403")) {
+        logMessage =
+            "\n\n Unable to find the application on TeamServer with the id ["
+                + applicationId
+                + "]\n";
+      } else {
+        logMessage =
+            "\n\n Unable to retrieve the application list from TeamServer. Please check that TeamServer is running at this address ["
+                + apiUrl
+                + "]\n";
+      }
+      throw new MojoExecutionException(logMessage, e);
+    }
+    if (applications.getApplication() == null) {
+      throw new MojoExecutionException(
+          "\n\nApplication with id '"
+              + applicationId
+              + "' not found. Make sure this application appears in TeamServer under the 'Applications' tab.\n");
+    }
+    return applications.getApplication().getName();
+  }
+
+  String computeAppVersion(Date currentDate) {
     if (computedAppVersion != null) {
       return computedAppVersion;
     }
@@ -121,7 +190,7 @@ public class InstallAgentContrastMavenMojo extends AbstractContrastMavenPluginMo
     return computedAppVersion;
   }
 
-  public String computeSessionMetadata() {
+  String computeSessionMetadata() {
     List<String> metadata = new ArrayList<String>();
 
     for (Map.Entry<String, String> entry : environmentToSessionMetadata.entrySet()) {
@@ -135,11 +204,11 @@ public class InstallAgentContrastMavenMojo extends AbstractContrastMavenPluginMo
     return StringUtils.join(metadata, ",");
   }
 
-  public String buildArgLine(String currentArgLine) {
+  String buildArgLine(String currentArgLine) {
     return buildArgLine(currentArgLine, appName);
   }
 
-  public String buildArgLine(String currentArgLine, String applicationName) {
+  String buildArgLine(String currentArgLine, String applicationName) {
 
     if (currentArgLine == null) {
       getLog().info("Current argLine is null");
@@ -206,4 +275,59 @@ public class InstallAgentContrastMavenMojo extends AbstractContrastMavenPluginMo
     getLog().info("Updated argLine is " + newArgLine);
     return newArgLine.trim();
   }
+
+  File installJavaAgent(ContrastSDK connection) throws MojoExecutionException {
+    byte[] javaAgent;
+    File agentFile;
+
+    if (StringUtils.isEmpty(jarPath)) {
+      getLog().info("No jar path was configured. Downloading the latest contrast.jar...");
+
+      try {
+        if (profile != null) {
+          javaAgent = connection.getAgent(AgentType.JAVA, orgUuid, profile);
+        } else {
+          javaAgent = connection.getAgent(AgentType.JAVA, orgUuid);
+        }
+      } catch (IOException e) {
+        throw new MojoExecutionException(
+            "\n\nWe couldn't download the Java agent from TeamServer with this user ["
+                + username
+                + "]. Please check that all your credentials are correct. If everything is correct, please contact Contrast Support. The error is:",
+            e);
+      } catch (UnauthorizedException e) {
+        throw new MojoExecutionException(
+            "\n\nWe contacted TeamServer successfully but couldn't authorize with the credentials you provided. The error is:",
+            e);
+      }
+
+      // Save the jar to the 'target' directory
+      agentFile = new File(project.getBuild().getDirectory() + File.separator + AGENT_NAME);
+
+      try {
+        FileUtils.writeByteArrayToFile(agentFile, javaAgent);
+      } catch (IOException e) {
+        throw new MojoExecutionException("Unable to save the latest java agent.", e);
+      }
+
+      getLog().info("Saved the latest java agent to " + agentFile.getAbsolutePath());
+      contrastAgentLocation = agentFile.getAbsolutePath();
+
+    } else {
+      getLog().info("Using configured jar path " + jarPath);
+
+      agentFile = new File(jarPath);
+
+      if (!agentFile.exists()) {
+        throw new MojoExecutionException("Unable to load the local Java agent from " + jarPath);
+      }
+
+      getLog().info("Loaded the latest java agent from " + jarPath);
+      contrastAgentLocation = jarPath;
+    }
+
+    return agentFile;
+  }
+
+  private static final String AGENT_NAME = "contrast.jar";
 }
