+++ src/main/java/com/contrastsecurity/maven/plugin/ContrastScanMojo.java
@@ -1,0 +1,7 @@
+package com.contrastsecurity.maven.plugin;
+
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
+
+@Mojo(name = "scan", defaultPhase = LifecyclePhase.INTEGRATION_TEST, requiresOnline = true)
+public class ContrastScanMojo extends AbstractContrastMojo {}
